BATTLE CONTROLLER REFACTORING
First commit of this branch, and the main thing it does is get the controller interacting more with the model. So the model calculates things that the controller then takes and acts upon. So far nothing really
occurs from that and it's not done, but once I finish it up and add everything I'll next need to add display methods.