BATTLE CONTROLLER REFACTORING 2: TESTER AND DEBUG, ACTUAL TURN RUNNING, INTERACT WITH THE MODEL 2
This commit adds more listening to the model for the battle controller and refactors a bit to make things cleaner. The structure of the battle controller is also coming into view and seems to be well-architected,
at least much better than my previous attempts. Methods are small and simple, and build into each other logically.

This also adds debug lines in the console for when an entity attacks, and there is definitely a bug in how the turn is being run, but for now that's okay. What I really need to do is go back to refactoring the
display and getting that to interact with the battle controller.

One thing that might be a path to go down, since right now I have the display panel and interact panel separate on the game frame, is that I pass in the game window/owner to the battle controller, and then use
that to access the display panels and such. I wonder if that would be a good idea....

Anyway I'll stay in this branch because it does involve the controller, but I might bring off the branch real quick to do that GUI and display refactoring.