BATTLE CONTROLLER REFACTORING 3: Adding Button Panel to Battle Controller + Comment refactoring
This commit adds the parameter of the button panel to the battle controller so that it can access its variables and the results from button presses. This commit also refactors comments to be more in line with
javadoc comments that I've seen elsewhere and make the comments less bulky and more understandable. There are more comments to be redone but for now this works to get the battle controller comments a bit better.
I'm learning from reading reconstructed RubyDungeon and Minecraft code, and I want to look at java projecs more and see what's up.