BATTLE CONTROLLER REFACTORING 4: Minor additions
This commit adds some minor additions to the controller and model classes, making them a bit better. For some reason I'm losing the plot of how things all interact, but it's nothing some refactoring later can't
solve. Now after this one genuinely I'm going to move to the display and refactor that and add proper functions. I might use the varargs printing methods I was using before but we'll see.