DISPLAY CONTROLLER INTERACT 2: PRINTING ATTACK
By adding return values to take damage function, I then bring the damage value through the entity attack method into the battle scene with attack entity which then allows the battle controller to access it and
print necessary data about the amount of damage done. This is the beginning of actually getting things together to display, and already the text log displays the attack of the entity. The only thing I need to do
now is reformat the log method so it immediately takes in a String.format() and uses varargs, and then also make it so that the stat displayers update when the attack is run.