DISPLAY CONTROLLER INTERACT (#17)
* DISPLAY CONTROLLER INTERACT

This commit begins the branch to start getting the display and controller to interact with one another in an actually productive way. This will first involve printing attacks, which I then want to refactor so
that my damn buttons actually do something.

* DISPLAY CONTROLLER INTERACT 2: PRINTING ATTACK

By adding return values to take damage function, I then bring the damage value through the entity attack method into the battle scene with attack entity which then allows the battle controller to access it and
print necessary data about the amount of damage done. This is the beginning of actually getting things together to display, and already the text log displays the attack of the entity. The only thing I need to do
now is reformat the log method so it immediately takes in a String.format() and uses varargs, and then also make it so that the stat displayers update when the attack is run.

* DISPLAY CONTROLLER INTERACT 3: Minor refactoring

This commit is a bit messed up and this branch is and the github shit so this is just to make things a bit more equal then I'm going to squash and merge this into the branch and start working on refactoring the
display itself and everything because I need to completely re-architect the flow of this game.