DISPLAY CONTROLLER INTERACT 4: buttons with attached action listeners
This commit attaches button listeners in the battle controller to the buttons of the GUI and adds the button initialization to the broader battle controller initialization. They are not actually implemented in
terms of doing anything in their action listeners other than attack, and more needs to be added, but this is a good step. Oh, and one thing that needs to happen is a slight bug fix; the action listener has to
take in two entities, but in the initialization step this means it will always use those same enemies as attacker and target. So what I need to do is set up a variable in the Battle State for who is currently
attacking, and use that to set the variables in the action listener.

This will be the next commit.