DISPLAY CONTROLLER INTERACT 6: WORKING PROTOTYPE (Mostly...)
This commit finishes refactoring and remaking the battle controller and is now WORKING!!! It properly does attacking from the player and attacking from the enemy. It also correctly does the who goes first calc
each turn, so sometimes per turn the slime goes first (if it's level 2) and sometimes the player goes first.

HOWEVER, there is a bug, and I think it's related to that; when the slime goes first after already going, I think the game gets stuck putting it to the enemy turn phase again, so the buttons get locked off b/c it
is not the player turn, but the game doesn't run the enemy attack because it doesn't seem to think it's the enemy's turn. This will stop the game in the middle and prevent any input, so I'll probably have to look
into that by printing the game phase and see how to fix it.

The battle controller also NEEDS to be refactored, but for right now it's okay. I'll do that refactoring for what I've created and it should be nice. But it's awesome, the stat displayers update and I might just
call an update to the entire GUI after every action... idk. But it WORKS for attacking and you can have a full battle!