DISPLAY CONTROLLER INTERACT 6: WORKING PROTOTYPE 2
This commit finishes with an ACTUAL GENUINE WORKING PROTOTYPE. There are certainly bugs that need to be fixed but functionality for all the buttons are added, and you can play a battle with the slime until either
the player or the slime dies. WOOHOO!!

Things to add for the future now:
REFACTORING! BUG FIXES!
Battle end codes, setting entities states when dead etc.
Tests for things like running
Action Battle Points system (will require a rewrite of battle logic)