BATTLE CONTROLLER REFACTORING (#16)
* BATTLE CONTROLLER REFACTORING

First commit of this branch, and the main thing it does is get the controller interacting more with the model. So the model calculates things that the controller then takes and acts upon. So far nothing really
occurs from that and it's not done, but once I finish it up and add everything I'll next need to add display methods.

* BATTLE CONTROLLER REFACTORING 2: TESTER AND DEBUG, ACTUAL TURN RUNNING, INTERACT WITH THE MODEL 2

This commit adds more listening to the model for the battle controller and refactors a bit to make things cleaner. The structure of the battle controller is also coming into view and seems to be well-architected,
at least much better than my previous attempts. Methods are small and simple, and build into each other logically.

This also adds debug lines in the console for when an entity attacks, and there is definitely a bug in how the turn is being run, but for now that's okay. What I really need to do is go back to refactoring the
display and getting that to interact with the battle controller.

One thing that might be a path to go down, since right now I have the display panel and interact panel separate on the game frame, is that I pass in the game window/owner to the battle controller, and then use
that to access the display panels and such. I wonder if that would be a good idea....

Anyway I'll stay in this branch because it does involve the controller, but I might bring off the branch real quick to do that GUI and display refactoring.

* BATTLE CONTROLLER REFACTORING 3: Adding Button Panel to Battle Controller + Comment refactoring

This commit adds the parameter of the button panel to the battle controller so that it can access its variables and the results from button presses. This commit also refactors comments to be more in line with
javadoc comments that I've seen elsewhere and make the comments less bulky and more understandable. There are more comments to be redone but for now this works to get the battle controller comments a bit better.
I'm learning from reading reconstructed RubyDungeon and Minecraft code, and I want to look at java projecs more and see what's up.

* BATTLE CONTROLLER REFACTORING 4: Minor additions

This commit adds some minor additions to the controller and model classes, making them a bit better. For some reason I'm losing the plot of how things all interact, but it's nothing some refactoring later can't
solve. Now after this one genuinely I'm going to move to the display and refactor that and add proper functions. I might use the varargs printing methods I was using before but we'll see.

* DISPLAY CONTROLLER INTERACT (#17)

* DISPLAY CONTROLLER INTERACT

This commit begins the branch to start getting the display and controller to interact with one another in an actually productive way. This will first involve printing attacks, which I then want to refactor so
that my damn buttons actually do something.

* DISPLAY CONTROLLER INTERACT 2: PRINTING ATTACK

By adding return values to take damage function, I then bring the damage value through the entity attack method into the battle scene with attack entity which then allows the battle controller to access it and
print necessary data about the amount of damage done. This is the beginning of actually getting things together to display, and already the text log displays the attack of the entity. The only thing I need to do
now is reformat the log method so it immediately takes in a String.format() and uses varargs, and then also make it so that the stat displayers update when the attack is run.

* DISPLAY CONTROLLER INTERACT 3: Minor refactoring

This commit is a bit messed up and this branch is and the github shit so this is just to make things a bit more equal then I'm going to squash and merge this into the branch and start working on refactoring the
display itself and everything because I need to completely re-architect the flow of this game.

* DISPLAY CONTROLLER INTERACT 4: buttons with attached action listeners

This commit attaches button listeners in the battle controller to the buttons of the GUI and adds the button initialization to the broader battle controller initialization. They are not actually implemented in
terms of doing anything in their action listeners other than attack, and more needs to be added, but this is a good step. Oh, and one thing that needs to happen is a slight bug fix; the action listener has to
take in two entities, but in the initialization step this means it will always use those same enemies as attacker and target. So what I need to do is set up a variable in the Battle State for who is currently
attacking, and use that to set the variables in the action listener.

This will be the next commit.

* DISPLAY CONTROLLER INTERACT 5: Battle buttons "working"

This commit gets the battle buttons actually working with action listeners applied to them. Right now those listeners don't do too much since it's never the player's turn (and hence player attack can never occur)
but their test debug action listener does work, printing out when they are pressed to the console. Now it's just a matter of getting the battle set up properly, the code refactored, and the buttons doing their
actual runcode.

* DISPLAY CONTROLLER INTERACT 6: WORKING PROTOTYPE (Mostly...)

This commit finishes refactoring and remaking the battle controller and is now WORKING!!! It properly does attacking from the player and attacking from the enemy. It also correctly does the who goes first calc
each turn, so sometimes per turn the slime goes first (if it's level 2) and sometimes the player goes first.

HOWEVER, there is a bug, and I think it's related to that; when the slime goes first after already going, I think the game gets stuck putting it to the enemy turn phase again, so the buttons get locked off b/c it
is not the player turn, but the game doesn't run the enemy attack because it doesn't seem to think it's the enemy's turn. This will stop the game in the middle and prevent any input, so I'll probably have to look
into that by printing the game phase and see how to fix it.

The battle controller also NEEDS to be refactored, but for right now it's okay. I'll do that refactoring for what I've created and it should be nice. But it's awesome, the stat displayers update and I might just
call an update to the entire GUI after every action... idk. But it WORKS for attacking and you can have a full battle!

* DISPLAY CONTROLLER INTERACT 6: WORKING PROTOTYPE 2

This commit finishes with an ACTUAL GENUINE WORKING PROTOTYPE. There are certainly bugs that need to be fixed but functionality for all the buttons are added, and you can play a battle with the slime until either
the player or the slime dies. WOOHOO!!

Things to add for the future now:
REFACTORING! BUG FIXES!
Battle end codes, setting entities states when dead etc.
Tests for things like running
Action Battle Points system (will require a rewrite of battle logic)