BATTLE PROGRESS REFACTORING
This branch deals with changing around the battle progress to make it better and work more effectively and error-proof. It involves creating a turn set object that holds the current go order of the turn set and uses that to act on it, so its essentially part of the controller class handling battle
progress and such.

However, it's a real royal mess. So I need to work on it and figure out how to get this thing to work. I've rewritten this code like 10 times over, and there's probably much smarter ways to do it, but I'm trying to figure this out basically all on my own.

My next thought process is to move the turn buttons into turn set as well (and rename it to turn controller) and so I can have both the player and the enemy turns end the turn set if needed rather than having this crisscross between it and the battle controller. But that also involves moving and
changing around the GUI to make sure there's the turn buttons panel and the general battle interact panel. So that's what will happen next.