BATTLE PROGRESS REFACTORING 2: Buttons MOVED
This commit successfully moves the buttons into the Turn Set object, so they are created whenever a new turn set is created. I don't know if that's particularly great, but oh well. The other thing is that it modified the tester to accommodate for this; refreshing the game window and adding the
button panel after every time the battle scene runs battle must be done.

Otherwise, this is working to the point where the player can click anything, but the enemy does not react. Now that must be done. But the player does remove themselves from the go Order, so I think all that must be done is running the run turn set again since it just calls getFirst. I need to add
that condition, but otherwise it will hopefully work?