GUI DETOUR
This branch is a gui detour related to the improvement of the battle system. What this does is it creates a base container panel that holds an interact and display panel. This will be used by the battle controller, so that is passed into it as a parameter, and then within the battle controller
the necessary components are added to the container panel. Now, the battle controller can swap these components within itself without having the game Window interfere.

This will allow the turn sets to have their own turn action button panel that they will give to the battle controller by taking in as a parameter the container panel, or by the battle controller doing it itself. This will allow the action listeners for turn buttons to be in the turn set, which will
allow for overall cleaner and better battle progression.

See? It is related!