GUI DETOUR 2: Container Panel
This commit adds the fully fledged container panel, which doesn't take any panels in but the panels can be set. It is added to the gameWindow constructor, which then adds it to the content pane, which allows other panels to be added to it with the setPanels method when calling it from the game
window object. Thus, with this, both battle panels are able to be added, and they properly show up when played. Now that these all accurately show up, it is time for the actual meat of getting the battle controller and turn sets to behave nicely and run the battle smoothly.

Unless there's anything in the GUI that crops up, then I'll deal with that.