GUI DETOUR (#19)
* GUI DETOUR

This branch is a gui detour related to the improvement of the battle system. What this does is it creates a base container panel that holds an interact and display panel. This will be used by the battle controller, so that is passed into it as a parameter, and then within the battle controller
the necessary components are added to the container panel. Now, the battle controller can swap these components within itself without having the game Window interfere.

This will allow the turn sets to have their own turn action button panel that they will give to the battle controller by taking in as a parameter the container panel, or by the battle controller doing it itself. This will allow the action listeners for turn buttons to be in the turn set, which will
allow for overall cleaner and better battle progression.

See? It is related!

* GUI DETOUR 2: Container Panel

This commit adds the fully fledged container panel, which doesn't take any panels in but the panels can be set. It is added to the gameWindow constructor, which then adds it to the content pane, which allows other panels to be added to it with the setPanels method when calling it from the game
window object. Thus, with this, both battle panels are able to be added, and they properly show up when played. Now that these all accurately show up, it is time for the actual meat of getting the battle controller and turn sets to behave nicely and run the battle smoothly.

Unless there's anything in the GUI that crops up, then I'll deal with that.

* GUI DETOUR 3: Finishing

This commit finishes this branch, even though it starts moving on to other things.