BATTLE PROGRESS REFACTORING 4(?): IT WORKS!!!
IT WORKS!!! The solution for this actually WORKED!! I put the progress methods in the player action listeners and added conditionals to run either the enemy's turn or the player's turn. This actually makes the turns work completely perfectly; even with higher level enemies; sometimes they attack
first, other times second, when speed is tied. It's all actually working.

There seem to be some UI bugs, but otherwise it is working. holy moly.

Next I need to add checks for death as well as better UI logging and display, and then I can move to catching errors and refactoring to make things make more sense and easier.

But I actually architected this, and it worked!!! I added proper conditionals and have an on End runnable as turn set runnable to start the next one and it works.