BATTLE PROGRESS REFACTORING 5: Re-org and comments, beginning of entity select
This commit re-organizes the code a bit and also adds the beginning for the new class, the entity select panel, which will be used when the player selects attack to determine which enemy the player wants to attack.

This will involve further refactoring of the turn set, and probably cause it to also take in the gui container as a parameter so it can also swap panels, but otherwise it shouldn't be too hard.