BATTLE PROGRESS REFACTORING (#18)
* BATTLE PROGRESS REFACTORING

This branch deals with changing around the battle progress to make it better and work more effectively and error-proof. It involves creating a turn set object that holds the current go order of the turn set and uses that to act on it, so its essentially part of the controller class handling battle
progress and such.

However, it's a real royal mess. So I need to work on it and figure out how to get this thing to work. I've rewritten this code like 10 times over, and there's probably much smarter ways to do it, but I'm trying to figure this out basically all on my own.

My next thought process is to move the turn buttons into turn set as well (and rename it to turn controller) and so I can have both the player and the enemy turns end the turn set if needed rather than having this crisscross between it and the battle controller. But that also involves moving and
changing around the GUI to make sure there's the turn buttons panel and the general battle interact panel. So that's what will happen next.

* BATTLE PROGRESS REFACTORING 2: Buttons MOVED

This commit successfully moves the buttons into the Turn Set object, so they are created whenever a new turn set is created. I don't know if that's particularly great, but oh well. The other thing is that it modified the tester to accommodate for this; refreshing the game window and adding the
button panel after every time the battle scene runs battle must be done.

Otherwise, this is working to the point where the player can click anything, but the enemy does not react. Now that must be done. But the player does remove themselves from the go Order, so I think all that must be done is running the run turn set again since it just calls getFirst. I need to add
that condition, but otherwise it will hopefully work?

* GUI DETOUR (#19)

* GUI DETOUR

This branch is a gui detour related to the improvement of the battle system. What this does is it creates a base container panel that holds an interact and display panel. This will be used by the battle controller, so that is passed into it as a parameter, and then within the battle controller
the necessary components are added to the container panel. Now, the battle controller can swap these components within itself without having the game Window interfere.

This will allow the turn sets to have their own turn action button panel that they will give to the battle controller by taking in as a parameter the container panel, or by the battle controller doing it itself. This will allow the action listeners for turn buttons to be in the turn set, which will
allow for overall cleaner and better battle progression.

See? It is related!

* GUI DETOUR 2: Container Panel

This commit adds the fully fledged container panel, which doesn't take any panels in but the panels can be set. It is added to the gameWindow constructor, which then adds it to the content pane, which allows other panels to be added to it with the setPanels method when calling it from the game
window object. Thus, with this, both battle panels are able to be added, and they properly show up when played. Now that these all accurately show up, it is time for the actual meat of getting the battle controller and turn sets to behave nicely and run the battle smoothly.

Unless there's anything in the GUI that crops up, then I'll deal with that.

* GUI DETOUR 3: Finishing

This commit finishes this branch, even though it starts moving on to other things.

* BATTLE PROGRESS REFACTORING 4(?): IT WORKS!!!

IT WORKS!!! The solution for this actually WORKED!! I put the progress methods in the player action listeners and added conditionals to run either the enemy's turn or the player's turn. This actually makes the turns work completely perfectly; even with higher level enemies; sometimes they attack
first, other times second, when speed is tied. It's all actually working.

There seem to be some UI bugs, but otherwise it is working. holy moly.

Next I need to add checks for death as well as better UI logging and display, and then I can move to catching errors and refactoring to make things make more sense and easier.

But I actually architected this, and it worked!!! I added proper conditionals and have an on End runnable as turn set runnable to start the next one and it works.

* BATTLE PROGRESS REFACTORING 5: Re-org and comments, beginning of entity select

This commit re-organizes the code a bit and also adds the beginning for the new class, the entity select panel, which will be used when the player selects attack to determine which enemy the player wants to attack.

This will involve further refactoring of the turn set, and probably cause it to also take in the gui container as a parameter so it can also swap panels, but otherwise it shouldn't be too hard.

* BATTLE PROGRESS REFACTORING 6: A SEA CHANGE

I am abandoning this branch (merging into main) because I must re-organize my code with interfaces. It would make things so much easier. So this branch is fine everything still works. But I'm moving on to a new branch now.