INTERFACES - GUI
This branch begins adding interfaces to the GUI elements to make things a lot easier and a lot better. This commit in particular adds the Writeable interface, which interacts with anything that will be writeable (in particular, text boxes). This can also include other things, however... maybe.
Okay maybe this wasn't the best use of an interface but it is something I wanted to try, and it makes things significantly nicer to look at.