INTERFACES - GUI 2: Text
This commit adds interfaces for text boxes and starts implementing them. They still need to be cleaned up, but already this is much better. It adds an interface that is implemented by the text display box class, which I should make abstract, and its children--main text box and stat displayer. This
will continue to be organized and made better.