INTERFACES - GUI 3: Displays
This commit finishes up the interface integration with displays, and all seems to be working. There may be some unforeseen bugs, but for now it all looks like its working; the only thing is to make sure all casts are in try and catch situations, but all that error handling will come later (as I
learn how to do it...)

Anyway, I'm going to move on to the interactables now.