INTERFACES - GUI 4: Interactables
This commit deals with adding interfaces for interactable components, named inputtables, as they deal with player input. There are more interfaces that can be added and things can be restructured, but this is at least a good start. With this, the GUI at least has the beginnings of interfaces,
and they do seem to be working properly upon launching the GUI.