INTERFACES - GUI (#20)
* INTERFACES - GUI

This branch begins adding interfaces to the GUI elements to make things a lot easier and a lot better. This commit in particular adds the Writeable interface, which interacts with anything that will be writeable (in particular, text boxes). This can also include other things, however... maybe.
Okay maybe this wasn't the best use of an interface but it is something I wanted to try, and it makes things significantly nicer to look at.

* INTERFACES - GUI 2: Text

This commit adds interfaces for text boxes and starts implementing them. They still need to be cleaned up, but already this is much better. It adds an interface that is implemented by the text display box class, which I should make abstract, and its children--main text box and stat displayer. This
will continue to be organized and made better.

* INTERFACES - GUI 3: Displays

This commit finishes up the interface integration with displays, and all seems to be working. There may be some unforeseen bugs, but for now it all looks like its working; the only thing is to make sure all casts are in try and catch situations, but all that error handling will come later (as I
learn how to do it...)

Anyway, I'm going to move on to the interactables now.

* INTERFACES - GUI 4: Interactables

This commit deals with adding interfaces for interactable components, named inputtables, as they deal with player input. There are more interfaces that can be added and things can be restructured, but this is at least a good start. With this, the GUI at least has the beginnings of interfaces,
and they do seem to be working properly upon launching the GUI.