BATTLE INTERFACE
This branch introduces interfaces into the battle structure, as well as reformatting it to be a bit better with how it divvies up responsibilities.

This adds a turn class, which stores a single entity turn and action. This can be refactored to just store the action rather than do anything with it, so the round can just store turn actions and then round controller deals with it.
Round class which holds the turns involved in that round.
Round Controller which runs these two model classes and calls upon TurnListener as an interface that the battle controller implements, so that the round controller can call upon these methods when certain things happen; i.e. a turn is executed, etc.

There was a bug where the turns seemed to lag one turn behind with all of their values. This was fixed by me realizing I was only defining the entity turn choice at turn instantiation, when I should have been defining it whenever runTurn was run. So entities were only running turns with the choice
they defaulted to at turn instantiation, which was either invalid or the one before.