BATTLE INTERFACE 2: BASIC FEATURE COMPLETE
This commit is almost the feature complete battle scene. It doesn't seem to have any bugs, but still does not have the item or run functionality.

However, it is using interfaces!

It can certainly be refactored with abstraction and flattening of some methods, but for now it works; there will be a refactoring branch later.